# swagger.yaml
openapi: 3.0.0
info:
    title: Job API
    description: A REST API for managing job listings. Users can create, read, update, and delete job listings.
    version: 1.0.0
    contact:
      name: API Support
      email: cauebrolesef@gmail.com
servers:
    - url: http://localhost:3000/api/v1
      description: Local Development Server
paths:
    /auth/register:
        post:
            tags:
                - Auth
            summary: Register User
            operationId: RegisterUser
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterUserRequest'
                        example:
                            name: bob
                            email: bob@gmail.com
                            password: secret
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
            security: []
    /auth/login:
        post:
            tags:
                - Auth
            summary: Login User
            operationId: LoginUser
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginUserRequest'
                        example:
                            email: susan@gmail.com
                            password: secret
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
            security: []
    /jobs:
        post:
            tags:
                - Jobs
            summary: Create Job
            operationId: CreateJob
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateJobRequest'
                        example:
                            company: google
                            position: front-end developer
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
        get:
            tags:
                - Jobs
            summary: Get All Jobs
            operationId: GetAllJobs
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
    /jobs/{id}:
        parameters:
            - in: path
              name: id
              schema:
                  type: string
              required: true
              description: The job ID
        get:
            tags:
                - Jobs
            summary: Get Single Job
            operationId: GetSingleJob
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false

        patch:
            tags:
                - Jobs
            summary: Update Job
            operationId: UpdateJob
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateJobRequest'
                        example:
                            company: random
                            position: intern
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
        delete:
            tags:
                - Jobs
            summary: Delete Job
            operationId: DeleteJob
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
components:
    schemas:
        RegisterUserRequest:
            title: RegisterUserRequest
            required:
                - name
                - email
                - password
            type: object
            properties:
                name:
                    type: string
                email:
                    type: string
                password:
                    type: string
            example:
                name: bob
                email: bob@gmail.com
                password: secret
        LoginUserRequest:
            title: LoginUserRequest
            required:
                - email
                - password
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
            example:
                email: susan@gmail.com
                password: secret
        CreateJobRequest:
            title: CreateJobRequest
            required:
                - company
                - position
            type: object
            properties:
                company:
                    type: string
                position:
                    type: string
            example:
                company: google
                position: front-end developer
        UpdateJobRequest:
            title: UpdateJobRequest
            required:
                - company
                - position
            type: object
            properties:
                company:
                    type: string
                position:
                    type: string
            example:
                company: random
                position: intern
    securitySchemes:
        httpBearer:
            type: http
            scheme: bearer
security:
    - httpBearer: []
tags:
    - name: Auth
      description: ''
    - name: Jobs
      description: ''